Moved the functionality in aec_core_internal.h into other files.
The purpose of this CL is to simplify upcoming AEC algorithm
changes.

The changes should be bitexact.

The presubmit was bypassed due to a presubmit complaint
about usage of short instead of int16_t which will be
addressed in upcoming CLs.

BUG=webrtc:5298, webrtc:5201

NOPRESUBMIT=true

Review-Url: https://codereview.webrtc.org/1949803004
Cr-Original-Commit-Position: refs/heads/master@{#12662}
Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: e687f7816c9e6b7e374d62a4d44ce4ee14a49168